+++ kompendium-core/src/main/kotlin/org/leafygreens/kompendium/Kompendium.kt
@@ -3,7 +3,6 @@ package org.leafygreens.kompendium
 import io.ktor.application.ApplicationCall
 import io.ktor.http.HttpMethod
 import io.ktor.routing.Route
-import io.ktor.routing.createRouteFromPath
 import io.ktor.routing.method
 import io.ktor.util.pipeline.PipelineInterceptor
 import java.lang.reflect.ParameterizedType
@@ -13,83 +12,128 @@ import kotlin.reflect.full.findAnnotation
 import kotlin.reflect.full.memberProperties
 import kotlin.reflect.jvm.javaField
 import org.leafygreens.kompendium.annotations.KompendiumField
-import org.leafygreens.kompendium.annotations.KompendiumInternal
+import org.leafygreens.kompendium.annotations.KompendiumRequest
+import org.leafygreens.kompendium.annotations.KompendiumResponse
+import org.leafygreens.kompendium.models.meta.MethodInfo
 import org.leafygreens.kompendium.models.oas.ArraySchema
 import org.leafygreens.kompendium.models.oas.FormatSchema
 import org.leafygreens.kompendium.models.oas.ObjectSchema
 import org.leafygreens.kompendium.models.oas.OpenApiSpec
 import org.leafygreens.kompendium.models.oas.OpenApiSpecComponentSchema
 import org.leafygreens.kompendium.models.oas.OpenApiSpecInfo
+import org.leafygreens.kompendium.models.oas.OpenApiSpecMediaType
 import org.leafygreens.kompendium.models.oas.OpenApiSpecPathItem
 import org.leafygreens.kompendium.models.oas.OpenApiSpecPathItemOperation
+import org.leafygreens.kompendium.models.oas.OpenApiSpecReferenceObject
+import org.leafygreens.kompendium.models.oas.OpenApiSpecRequest
+import org.leafygreens.kompendium.models.oas.OpenApiSpecResponse
 import org.leafygreens.kompendium.models.oas.SimpleSchema
-import org.leafygreens.kompendium.models.meta.MethodInfo
 import org.leafygreens.kompendium.util.Helpers.calculatePath
 import org.leafygreens.kompendium.util.Helpers.putPairIfAbsent
 
 object Kompendium {
-  val openApiSpec = OpenApiSpec(
+
+  const val COMPONENT_SLUG = "#/components/schemas"
+
+  var openApiSpec = OpenApiSpec(
     info = OpenApiSpecInfo(),
     servers = mutableListOf(),
     paths = mutableMapOf()
   )
 
-  fun Route.notarizedGet(info: MethodInfo, body: PipelineInterceptor<Unit, ApplicationCall>): Route {
+  inline fun <reified TParam : Any, reified TResp : Any> Route.notarizedGet(
+    info: MethodInfo,
+    noinline body: PipelineInterceptor<Unit, ApplicationCall>
+  ): Route = generateComponentSchemas<TParam, Unit, TResp>() {
     val path = calculatePath()
     openApiSpec.paths.getOrPut(path) { OpenApiSpecPathItem() }
-    openApiSpec.paths[path]?.get = OpenApiSpecPathItemOperation(
-      summary = info.summary,
-      description = info.description,
-      tags = info.tags
-    )
+    openApiSpec.paths[path]?.get = info.parseMethodInfo<Unit, TResp>()
     return method(HttpMethod.Get) { handle(body) }
   }
 
-  inline fun <reified TQ : Any, reified TP : Any, reified TR : Any> Route.notarizedPost(
+  inline fun <reified TParam : Any, reified TReq : Any, reified TResp : Any> Route.notarizedPost(
     info: MethodInfo,
     noinline body: PipelineInterceptor<Unit, ApplicationCall>
-  ): Route = generateComponentSchemas<TQ, TP, TR>(info, body) { i, b ->
+  ): Route = generateComponentSchemas<TParam, TReq, TResp>() {
     val path = calculatePath()
     openApiSpec.paths.getOrPut(path) { OpenApiSpecPathItem() }
-    openApiSpec.paths[path]?.post = OpenApiSpecPathItemOperation(
-      summary = i.summary,
-      description = i.description,
-      tags = i.tags
-    )
-    return method(HttpMethod.Post) { handle(b) }
+    openApiSpec.paths[path]?.post = info.parseMethodInfo<TReq, TResp>()
+    return method(HttpMethod.Post) { handle(body) }
   }
 
-  inline fun <reified TQ : Any, reified TP : Any, reified TR : Any> Route.notarizedPut(
+  inline fun <reified TParam : Any, reified TReq : Any, reified TResp : Any> Route.notarizedPut(
     info: MethodInfo,
     noinline body: PipelineInterceptor<Unit, ApplicationCall>,
-  ): Route = generateComponentSchemas<TQ, TP, TR>(info, body) { i, b ->
+  ): Route = generateComponentSchemas<TParam, TReq, TResp>() {
     val path = calculatePath()
     openApiSpec.paths.getOrPut(path) { OpenApiSpecPathItem() }
-    openApiSpec.paths[path]?.put = OpenApiSpecPathItemOperation(
-      summary = i.summary,
-      description = i.description,
-      tags = i.tags
-    )
-    return method(HttpMethod.Put) { handle(b) }
+    openApiSpec.paths[path]?.put = info.parseMethodInfo<TReq, TResp>()
+    return method(HttpMethod.Put) { handle(body) }
   }
 
-  @OptIn(KompendiumInternal::class)
-  inline fun <reified TQ : Any, reified TP : Any, reified TR : Any> generateComponentSchemas(
+  inline fun <reified TParam : Any, reified TResp : Any> Route.notarizedDelete(
     info: MethodInfo,
-    noinline body: PipelineInterceptor<Unit, ApplicationCall>,
-    block: (MethodInfo, PipelineInterceptor<Unit, ApplicationCall>) -> Route
+    noinline body: PipelineInterceptor<Unit, ApplicationCall>
+  ): Route = generateComponentSchemas<TParam, Unit, TResp> {
+    val path = calculatePath()
+    openApiSpec.paths.getOrPut(path) { OpenApiSpecPathItem() }
+    openApiSpec.paths[path]?.delete = info.parseMethodInfo<Unit, TResp>()
+    return method(HttpMethod.Delete) { handle(body) }
+  }
+
+  inline fun <reified TReq, reified TResp> MethodInfo.parseMethodInfo() = OpenApiSpecPathItemOperation(
+    summary = this.summary,
+    description = this.description,
+    tags = this.tags,
+    deprecated = this.deprecated,
+    responses = parseResponseAnnotation<TResp>()?.let { mapOf(it) },
+    requestBody = parseRequestAnnotation<TReq>()
+  )
+
+  inline fun <reified TParam : Any, reified TReq : Any, reified TResp : Any> generateComponentSchemas(
+    block: () -> Route
   ): Route {
-    openApiSpec.components.schemas.putPairIfAbsent(objectSchemaPair(TQ::class))
-    openApiSpec.components.schemas.putPairIfAbsent(objectSchemaPair(TR::class))
-    openApiSpec.components.schemas.putPairIfAbsent(objectSchemaPair(TP::class))
-    return block.invoke(info, body)
+    if (TResp::class != Unit::class) openApiSpec.components.schemas.putPairIfAbsent(objectSchemaPair(TResp::class))
+    if (TReq::class != Unit::class) openApiSpec.components.schemas.putPairIfAbsent(objectSchemaPair(TReq::class))
+//    openApiSpec.components.schemas.putPairIfAbsent(objectSchemaPair(TParam::class))
+    return block.invoke()
   }
 
-  @KompendiumInternal
+  inline fun <reified TReq> parseRequestAnnotation(): OpenApiSpecRequest? = when (TReq::class) {
+    Unit::class -> null
+    else -> {
+      val anny = TReq::class.findAnnotation<KompendiumRequest>() ?: error("My way or the highway bub")
+      OpenApiSpecRequest(
+        description = anny.description,
+        content = anny.mediaTypes.associate {
+          val ref = OpenApiSpecReferenceObject("$COMPONENT_SLUG/${TReq::class.simpleName}")
+          val mediaType = OpenApiSpecMediaType.Referenced(ref)
+          Pair(it, mediaType)
+        }
+      )
+    }
+  }
+
+  inline fun <reified TResp> parseResponseAnnotation(): Pair<Int, OpenApiSpecResponse>? = when (TResp::class) {
+    Unit::class -> null
+    else -> {
+      val anny = TResp::class.findAnnotation<KompendiumResponse>() ?: error("My way or the highway bub")
+      val specResponse = OpenApiSpecResponse(
+        description = anny.description,
+        content = anny.mediaTypes.associate {
+          val ref = OpenApiSpecReferenceObject("$COMPONENT_SLUG/${TResp::class.simpleName}")
+          val mediaType = OpenApiSpecMediaType.Referenced(ref)
+          Pair(it, mediaType)
+        }
+      )
+      Pair(anny.status, specResponse)
+    }
+  }
+
   // TODO Investigate a caching mechanism to reduce overhead... then just reference once created
   fun objectSchemaPair(clazz: KClass<*>): Pair<String, ObjectSchema> {
     val o = objectSchema(clazz)
-    return Pair(clazz.qualifiedName!!, o)
+    return Pair(clazz.simpleName!!, o)
   }
 
   private fun objectSchema(clazz: KClass<*>): ObjectSchema =
@@ -115,7 +159,6 @@ object Kompendium {
     return ArraySchema(fieldToSchema(listType))
   }
 
-  @OptIn(KompendiumInternal::class)
   private fun fieldToSchema(field: KClass<*>): OpenApiSpecComponentSchema = when (field) {
     Int::class -> FormatSchema("int32", "integer")
     Long::class -> FormatSchema("int64", "integer")
@@ -124,5 +167,13 @@ object Kompendium {
     String::class -> SimpleSchema("string")
     Boolean::class -> SimpleSchema("boolean")
     else -> objectSchema(field)
+  }
+
+  internal fun resetSchema() {
+    openApiSpec = OpenApiSpec(
+      info = OpenApiSpecInfo(),
+      servers = mutableListOf(),
+      paths = mutableMapOf()
+    )
   }
 }