+++ kompendium-core/src/main/kotlin/org/leafygreens/kompendium/annotations/KompendiumModule.kt
@@ -1,5 +1,0 @@
-package org.leafygreens.kompendium.annotations
-
-@Retention(AnnotationRetention.SOURCE)
-@Target(AnnotationTarget.FUNCTION)
-annotation class KompendiumModule