+++ kompendium-core/src/main/kotlin/org/leafygreens/kompendium/annotations/KompendiumServers.kt
@@ -1,7 +1,0 @@
-package org.leafygreens.kompendium.annotations
-
-@Retention(AnnotationRetention.SOURCE)
-@Target(AnnotationTarget.FUNCTION, AnnotationTarget.CLASS)
-annotation class KompendiumServers(
-  val urls: Array<String>
-)