+++ kompendium-core/src/main/kotlin/org/leafygreens/kompendium/models/meta/MethodInfo.kt
@@ -1,3 +1,8 @@
 package org.leafygreens.kompendium.models.meta
 
-data class MethodInfo(val summary: String, val description: String? = null, val tags: Set<String> = emptySet())
+data class MethodInfo(
+  val summary: String,
+  val description: String? = null,
+  val tags: Set<String> = emptySet(),
+  val deprecated: Boolean = false
+)